requirements-development.txt
pytest
pytest-cov
pytest-django
pytest-mock
pytest-pythonpath
pytest-splinter
pytest-sugar

-r requirements.txt
